Fix case where dist being built is the dist that contains the share dir
When building the a dist that matches the name of the dist which has the share
dir, we look in $self->zilla->root for the file rather than in an installed
share dir. This fixes an issue where a bundle that uses this plugin cannot be
built until the bundle is installed, but you can't install it because you
can't build it.